♻️ extract ability types into shared lib
Ref: #152
