+++ libs/ability/.eslintrc.cjs
@@ -1,0 +1,13 @@
+module.exports = {
+  root: true,
+  extends: '@os',
+  parserOptions: {
+    tsconfigRootDir: __dirname,
+  },
+  overrides: [
+    {
+      files: 'src/',
+      extends: 'noise/node',
+    },
+  ],
+};
