+++ libs/ability/package.json
@@ -1,0 +1,36 @@
+{
+  "name": "@os/ability",
+  "version": "0.0.1",
+  "private": true,
+  "license": "AGPL-3.0",
+  "exports": {
+    ".": {
+      "import": "./dist/index.mjs",
+      "require": "./dist/index.js"
+    }
+  },
+  "engines": {
+    "node": "16.15.0"
+  },
+  "scripts": {
+    "build": "rollup -c",
+    "clean": "rm -rf .eslintcache .turbo/ dist/",
+    "dev": "rollup -c -w",
+    "lint": "eslint --cache --ext .js,.ts ."
+  },
+  "files": [
+    "dist/"
+  ],
+  "dependencies": {
+    "@casl/ability": "^5.4.3"
+  },
+  "devDependencies": {
+    "@types/eslint": "^8",
+    "eslint": "^8.15.0",
+    "rollup": "^2.72.0",
+    "rollup-plugin-dts": "^4.2.1",
+    "rollup-plugin-esbuild": "^4.9.1",
+    "typescript": "~4.6.4"
+  },
+  "types": "dist/index.d.ts"
+}
