+++ libs/ability/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "$schema": "https://json.schemastore.org/tsconfig",
+  "extends": "@os/tsconfig/lib.json",
+  "display": "ability",
+  "compilerOptions": {
+    "baseUrl": "."
+  },
+  "include": ["src"]
+}
